requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.0.0
pyotp==2.8.0
qrcode==7.4.2
cryptography==41.0.5
python-dotenv==1.0.0
Pillow==10.0.0
PyJWT==2.8.0
sympy==1.12
